Add parsing API
Add the API for parsing various representations of the Architecture-as-Code
language specification. Since we're using YAML in the official AaC repository,
go with YAML, for now, but this will allow parsing different representations
without having to touch the API.

Also add a helper macro that lets the user execute some code with *aac-spec*
bound to the specification read from the provided file.