Initial version of loading the spec
This is the initial crappy version of building the spec. What this means is that
things aren't as clean as I'd like them to be but it's loading everything from
the yaml spec file and creating a list of data and enum objects representing the
spec that was read in.